refactor(geo): move scan ZDR policy and TTL cache into dedicated modules
